--- a/x/mongo/driver/driver.go
+++ b/x/mongo/driver/driver.go
@@ -61,7 +61,7 @@ type Server interface {
 // Connection represents a connection to a MongoDB server.
 type Connection interface {
 	WriteWireMessage(context.Context, []byte) error
-	ReadWireMessage(ctx context.Context, dst []byte) ([]byte, error)
+	ReadWireMessage(ctx context.Context, dst *[]byte) ([]byte, error)
 	Description() description.Server
 
 	// Close closes any underlying connection and returns or frees any resources held by the