--- a/x/mongo/driver/topology/connection.go
+++ b/x/mongo/driver/topology/connection.go
@@ -383,9 +383,9 @@ func (c *connection) write(ctx context.Context, wm []byte) (err error) {
 }
 
 // readWireMessage reads a wiremessage from the connection. The dst parameter will be overwritten.
-func (c *connection) readWireMessage(ctx context.Context, dst []byte) ([]byte, error) {
+func (c *connection) readWireMessage(ctx context.Context, dst *[]byte) ([]byte, error) {
 	if atomic.LoadInt64(&c.state) != connConnected {
-		return dst, ConnectionError{ConnectionID: c.id, message: "connection is closed"}
+		return nil, ConnectionError{ConnectionID: c.id, message: "connection is closed"}
 	}
 
 	select {
@@ -411,7 +411,7 @@ func (c *connection) readWireMessage(ctx context.Context, dst []byte) ([]byte, e
 		return nil, ConnectionError{ConnectionID: c.id, Wrapped: err, message: "failed to set read deadline"}
 	}
 
-	dst, errMsg, err := c.read(ctx, dst)
+	res, errMsg, err := c.read(ctx, dst)
 	if err != nil {
 		// We closeConnection the connection because we don't know if there are other bytes left to read.
 		c.close()
@@ -426,10 +426,10 @@ func (c *connection) readWireMessage(ctx context.Context, dst []byte) ([]byte, e
 		}
 	}
 
-	return dst, nil
+	return res, nil
 }
 
-func (c *connection) read(ctx context.Context, dst []byte) (bytesRead []byte, errMsg string, err error) {
+func (c *connection) read(ctx context.Context, dst *[]byte) (bytesRead []byte, errMsg string, err error) {
 	go c.cancellationListener.Listen(ctx, c.cancellationListenerCallback)
 	defer func() {
 		// If the context is cancelled after we finish reading the server response, the cancellation listener could fire
@@ -467,21 +467,21 @@ func (c *connection) read(ctx context.Context, dst []byte) (bytesRead []byte, er
 		return nil, errResponseTooLarge.Error(), errResponseTooLarge
 	}
 
-	if int(size) > cap(dst) {
+	if dst == nil || int(size) > cap(*dst) {
 		// Since we can't grow this slice without allocating, just allocate an entirely new slice.
-		dst = make([]byte, 0, size)
+		*dst = make([]byte, 0, size)
 	}
 	// We need to ensure we don't accidentally read into a subsequent wire message, so we set the
 	// size to read exactly this wire message.
-	dst = dst[:size]
-	copy(dst, sizeBuf[:])
+	*dst = (*dst)[:size]
+	copy(*dst, sizeBuf[:])
 
-	_, err = io.ReadFull(c.nc, dst[4:])
+	_, err = io.ReadFull(c.nc, (*dst)[4:])
 	if err != nil {
 		return nil, "incomplete read of full message", err
 	}
 
-	return dst, "", nil
+	return *dst, "", nil
 }
 
 func (c *connection) close() error {
@@ -576,7 +576,7 @@ func (c initConnection) LocalAddress() address.Address {
 func (c initConnection) WriteWireMessage(ctx context.Context, wm []byte) error {
 	return c.writeWireMessage(ctx, wm)
 }
-func (c initConnection) ReadWireMessage(ctx context.Context, dst []byte) ([]byte, error) {
+func (c initConnection) ReadWireMessage(ctx context.Context, dst *[]byte) ([]byte, error) {
 	return c.readWireMessage(ctx, dst)
 }
 func (c initConnection) SetStreaming(streaming bool) {
@@ -619,11 +619,11 @@ func (c *Connection) WriteWireMessage(ctx context.Context, wm []byte) error {
 
 // ReadWireMessage handles reading a wire message from the underlying connection. The dst parameter
 // will be overwritten with the new wire message.
-func (c *Connection) ReadWireMessage(ctx context.Context, dst []byte) ([]byte, error) {
+func (c *Connection) ReadWireMessage(ctx context.Context, dst *[]byte) ([]byte, error) {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
 	if c.connection == nil {
-		return dst, ErrConnectionClosed
+		return nil, ErrConnectionClosed
 	}
 	return c.readWireMessage(ctx, dst)
 }