[GODRIVER-1670](https://jira.mongodb.org/browse/GODRIVER-1670)
benchmark/operaton_test.go shows the first [commit](https://github.com/mongodb/mongo-go-driver/commit/b23a13b4182a80f7e6f9eff7484adb316de167e8) achieves:
```
name                   old time/op    new time/op    delta
Operation/benchmark-4    93.4µs ± 3%    91.0µs ± 4%   -2.57%  (p=0.006 n=9+10)

name                   old alloc/op   new alloc/op   delta
Operation/benchmark-4    1.59kB ± 0%    1.50kB ± 0%   -6.03%  (p=0.000 n=9+8)

name                   old allocs/op  new allocs/op  delta
Operation/benchmark-4      30.0 ± 0%      27.0 ± 0%  -10.00%  (p=0.000 n=10+10)
```
The benchmark was running against a local live service, so the op time may have noise (e.g. network latency). However, the memory usage and allocation are relatively accurate.

TODOs:
- [ ] Adapt the interface from raw slice to io.Reader/io.Writer for better encapsulation.
- [ ] Use a streaming manner to consume the wire message to avoid additional memory copies.