@qingyang-hu 2022-07-25
I'm going to close this PR because we plan to refactor the interfaces to resolve the memory allocation issue.